+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 available on:
  ▶️ https://www.npmjs.com/package/@chat21/chat21-server
 
+## v0.2.33 - online
+- presence webhook and observer.webhooks...presence introduced
+- updated chat21client.js => v0.1.12.4 with 'presence' publish on.connect()
+
 ## v0.2.32 - online
 - RESTORED if (savedMessage.attributes && savedMessage.attributes.updateconversation == false) {update_conversation = false}. See v0.2.26